requirements.txt
# Python-Vi-ArrowKeys requirements

#keyboard
pynput
wheel
pillow
pystray